@@ -3,7 +3,7 @@ class RolesController < ApplicationController
   skip_before_action :connect_to_db, except: %i(new edit)
 
   def index
-    @roles = current_account.roles.order(:name).to_a if current_account.enterprise?
+    @roles = current_account.roles.order(:name).to_a
   end
 
   def new
@@ -18,7 +18,6 @@ class RolesController < ApplicationController
   def create
     @role = current_account.roles.build role_params
     if @role.save
-      log_events
       redirect_to roles_url, success: 'Role successfully created'
     else
       render :new
@@ -28,7 +27,6 @@ class RolesController < ApplicationController
   def update
     @role = current_account.roles.find params[:id]
     if @role.update role_params
-      log_events
       redirect_to roles_url, success: 'Role successfully updated'
     else
       render :new
@@ -38,23 +36,10 @@ class RolesController < ApplicationController
   def destroy
     @role = current_account.roles.find params[:id]
     @role.destroy!
-    log_events
     redirect_to roles_url, success: 'Role successfully destroyed'
   end
 
   private
-
-    def log_events
-      base = { account: current_account.name, username: current_user&.name, role: "#{@role.name} (id:#{@role.id})" }
-      return logger.warn base.merge(action: 'destroy-role').to_json if action_name == 'destroy'
-      if @role.saved_change_to_permissions?
-        logger.warn base.merge(action: 'update-role-permissions',
-          permissions: Hashdiff.diff(*@role.saved_changes[:permissions])).to_json
-      end
-      if role_params[:collaborator_ids].present?
-        logger.warn base.merge(action: 'update-role-members', collaborators: role_params[:collaborator_ids])
-      end
-    end
 
     def role_params
       params.require(:role).permit(:name, collaborator_ids: [], permissions: {})
