@@ -1,7 +1,5 @@
 class Account < ApplicationRecord
-  serialize :plan_migrations
-  before_create :setup_api_key
-  before_save :fill_adapter, :track_plan_migration
+  before_save :fill_adapter
   has_many :collaborators, dependent: :destroy
   has_many :users, through: :collaborators
   has_many :roles, dependent: :destroy
@@ -10,7 +8,6 @@ class Account < ApplicationRecord
   has_many :time_chart_widgets, dependent: :destroy
   has_many :pie_chart_widgets, dependent: :destroy
   has_many :stat_chart_widgets, dependent: :destroy
-  has_many :sign_ons, dependent: :destroy
   has_many :searches, dependent: :destroy
   has_many :table_configurations, dependent: :destroy
 
@@ -24,70 +21,11 @@ class Account < ApplicationRecord
   attr_encrypted :db_url, key: Rails.application.secrets.encryption_key, algorithm: 'aes-256-cbc',
                           v2_gcm_iv: true, mode: :per_attribute_iv_and_salt
 
-  scope :deleted, -> {where plan: Plan::DELETED}
-  scope :not_deleted, -> {where.not plan: Plan::DELETED}
 
   TIPS = %w(basic_search editing enumerable export_import displayed_record advanced_search serialized relationships time_charts keyboard_shortcuts time_zones).freeze
 
-  class Plan
-    PET_PROJECT = 'petproject'.freeze
-    STARTUP = 'startup'.freeze
-    ENTERPRISE = 'enterprise'.freeze
-    COMPLIMENTARY = 'complimentary'.freeze
-    DELETED = 'deleted'.freeze
-  end
-
-  def to_param
-    api_key
-  end
-
-  def self.fetch_missing_owner_emails
-    where(owner_email: nil).where.not(callback_url: nil).find_each do |account|
-      res = account.fetch_info
-      next unless res
-      account.update owner_email: res['owner_email']
-      account.update name: res['name'] unless account.name?
-    end
-  end
-
-  def fetch_info
-    auth = [HEROKU_MANIFEST['id'], HEROKU_MANIFEST['api']['password']].join(':')
-    res = RestClient.get "https://#{auth}@api.heroku.com/vendor/apps/#{callback_url.split('/').last.strip}"
-    JSON.parse res
-  rescue RestClient::ResourceNotFound
-    {}
-  end
-
   def valid_db_url?
     db_url.present?
-  end
-
-  def pet_project?
-    plan == Plan::PET_PROJECT
-  end
-
-  def startup?
-    plan == Plan::STARTUP
-  end
-
-  def free_plan?
-    pet_project? || complimentary?
-  end
-
-  def enterprise?
-    (plan == Plan::ENTERPRISE) || complimentary?
-  end
-
-  def complimentary?
-    plan == Plan::COMPLIMENTARY
-  end
-
-  def upgrade_link
-    'https://addons.heroku.com/adminium'
-  end
-
-  def flag_as_deleted!
-    update! db_url: nil, api_key: nil, plan: Plan::DELETED, deleted_at: Time.current
   end
 
   def displayed_next_tip
@@ -99,23 +37,11 @@ class Account < ApplicationRecord
     tip
   end
 
-  def reactivate attributes
-    update attributes.merge(deleted_at: nil, api_key: generate_api_key)
-  end
-
   def table_configuration_for table
     table_configurations.find_or_create_by! table: table
   end
 
   private
-
-  def setup_api_key
-    self.api_key = generate_api_key
-  end
-
-  def generate_api_key
-    SecureRandom.hex[0..8]
-  end
 
   def db_url_validation
     return if db_url.blank? || errors[:db_url].any?
@@ -128,11 +54,5 @@ class Account < ApplicationRecord
 
   def fill_adapter
     self.adapter = db_url.split(':').first if db_url? && encrypted_db_url_changed?
-  end
-
-  def track_plan_migration
-    self.plan_migrations ||= []
-    last_plan = plan_migrations.last[:plan] if plan_migrations.last
-    plan_migrations.push migrated_at: Time.current, plan: plan if last_plan.nil? || last_plan != plan
   end
 end
