@@ -3,10 +3,8 @@ class Collaborator < ApplicationRecord
   belongs_to :account
   validates :email, presence: true,
                     format: {with: /([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})/i}
-  before_create :match_with_existing_user
-  after_create :mail_collaborator
   has_and_belongs_to_many :roles
-  attr_accessor :domain
+  before_create :generate_token
 
   # FIXME: check table existence
   def permissions
@@ -23,11 +21,8 @@ class Collaborator < ApplicationRecord
     user&.name || email
   end
 
-  def match_with_existing_user
-    self.user = User.find_by email: email, provider: kind unless user
-  end
-
-  def mail_collaborator
-    CollaboratorMailer.notify_collaboration(self, domain).deliver_later if kind == 'google_oauth2'
+  private
+  def generate_token
+    self.token = SecureRandom.urlsafe_base64(16)
   end
 end
