@@ -2,7 +2,7 @@ class Role < ApplicationRecord
   validates :name, presence: true
   belongs_to :account
   has_and_belongs_to_many :collaborators
-  serialize :permissions
+  serialize :permissions, JSON
   before_save :normalize_permissions
 
   def to_s
