@@ -1,12 +1,10 @@
 FactoryBot.define do
   factory :account do
-    owner_email { 'john.doe@email.com' }
     db_url { Rails.configuration.test_database_conn_spec }
-    plan { Account::Plan::STARTUP }
   end
 
   factory :user do
-    email { 'john.doe@emai.com' }
+    sequence(:email) { |n| "joe#{n}@mail.com" }
   end
 
   factory :collaborator do
