@@ -137,6 +137,7 @@ class ChartsTest < ActionDispatch::IntegrationTest
     def display_chart table, column, svg: true
       visit resources_path(table) if table
       find("th[data-column-name=\"#{column}\"] i.time-chart").click
-      assert_selector '#chart_div svg' if svg
+      return unless svg
+      Capybara.using_wait_time(5) { assert_selector '#chart_div svg' }
     end
 end
